requirements.txt
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python>=4.8.0.76
numpy>=1.24.0
scikit-learn>=1.3.0
hdbscan>=0.8.33
scikit-image>=0.22.0
matplotlib>=3.7.2
pandas>=2.0.3
tqdm>=4.66.1
Pillow>=9.5.0
# Optional: BRISQUE (will be used if available)
pybrisque>=1.0
